pyproject.toml
[tool.poetry]
name = "frappe_library"
version = "1.0"
description = "Library system for Frappe"
authors = ["Anvesh Mishra <anveshmishra0508@gmail.com>"]
maintainers = [
    "Anvesh Mishra <anveshmishra0508@gmail.com>",
]
repository = "https://github.com/Anv3sh/frappe_library"
readme = "../../README.md"
keywords = ["library","system"]
packages = [{ include = "frappe_library", from = "./" }]
include = ["frappe_library/*", "frappe_library/**/*"] 
  
[tool.poetry.dependencies]  
python = "^3.9"  
annotated-types = "^0.6.0"  
blinker = "^1.7.0"  
build = "^1.1.1"  
CacheControl = "^0.14.0"  
certifi = "^2024.2.2"  
cffi = "^1.16.0"  
charset-normalizer = "^3.3.2"  
cleo = "^2.1.0"  
click = "^8.1.7"  
crashtest = "^0.4.1"  
distlib = "^0.3.8"  
dulwich = "^0.21.7"  
fastjsonschema = "^2.19.1"  
filelock = "^3.13.1"  
Flask = "^3.0.2"  
gunicorn = "^21.2.0"  
idna = "^3.6"  
importlib_metadata = "^7.0.2"  
installer = "^0.7.0"  
itsdangerous = "^2.1.2"  
Jinja2 = "^3.1.3"  
keyring = "^24.3.1"  
MarkupSafe = "^2.1.5"  
more-itertools = "^10.2.0"  
msgpack = "^1.0.8"  
packaging = "^23.2"  
pexpect = "^4.9.0"  
pkginfo = "^1.10.0"  
platformdirs = "^4.2.0"  
poetry = "^1.8.2"  
poetry-core = "^1.9.0"  
poetry-plugin-export = "^1.6.0"  
ptyprocess = "^0.7.0"  
pycparser = "^2.21"  
pydantic = "^2.6.3"  
pydantic_core = "^2.16.3"  
pyproject_hooks = "^1.0.0"  
rapidfuzz = "^3.6.2"  
requests = "^2.31.0"  
requests-toolbelt = "^1.0.0"  
shellingham = "^1.5.4"  
SQLAlchemy = "^2.0.28"  
sqlmodel = "^0.0.16"  
tomli = "^2.0.1"  
tomlkit = "^0.12.4"  
trove-classifiers = "^2024.3.3"  
typing_extensions = "^4.10.0"  
urllib3 = "^2.2.1"  
virtualenv = "^20.25.1"  
Werkzeug = "^3.0.1"  
xattr = "^1.1.0"  
zipp = "^3.17.0"
orjson = "^3.9.15"
flask_restful = "^0.3.10"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
typing = "^3.7.4.3"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.27.1"
dnspython = "^2.6.1"
email-validator = "^2.1.1"
flask-cors = "^4.0.0"
 

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.2"
mypy = "^1.1.1"
ruff = "^0.0.254"
httpx = "*"
pytest = "^7.2.2"
types-requests = "^2.28.11"
requests = "^2.28.0"
pytest-cov = "^4.0.0"
types-appdirs = "^1.4.3.5"
types-pyyaml = "^6.0.12.8"
jinja2 = "^3.1.2 "
  
[build-system]  
requires = ["poetry-core>=1.0.0"]  
build-backend = "poetry.core.masonry.api"  
